Add function to show monitor identity
