+++ Source/Monitorian.Core/Views/MainWindow.xaml
@@ -496,14 +496,14 @@
 									Background="Transparent">
 								<Grid Margin="{TemplateBinding Padding}">
 									<Grid Height="{Binding RelativeSource={RelativeSource Self}, Path=ActualWidth}"
-										  Margin="7,6" VerticalAlignment="Center"
-										  RenderOptions.EdgeMode="Aliased">
+											  Margin="7,6" VerticalAlignment="Center"
+											  RenderOptions.EdgeMode="Aliased">
 										<Path x:Name="StrokeIcon"
-											  Data="{TemplateBinding StrokeIconData}" Stretch="Uniform"
-											  Stroke="{TemplateBinding Foreground}" StrokeThickness="1"/>
+												  Data="{TemplateBinding StrokeIconData}" Stretch="Uniform"
+												  Stroke="{TemplateBinding Foreground}" StrokeThickness="1"/>
 										<Path x:Name="FillIcon"
-											  Data="{TemplateBinding FillIconData}" Stretch="Uniform"
-											  Fill="{TemplateBinding Foreground}"/>
+												  Data="{TemplateBinding FillIconData}" Stretch="Uniform"
+												  Fill="{TemplateBinding Foreground}"/>
 									</Grid>
 								</Grid>
 							</Border>
@@ -578,6 +578,11 @@
 			<Setter Property="FillIconData" Value="M 0,0 L 0,10 5,10 5,0 Z M 10,0 L 10,10"/>
 		</Style>
 
+		<Style x:Key="IdentityButtonStyle" TargetType="{x:Type controls:IconButton}" BasedOn="{StaticResource IconButtonStyle}">
+			<Setter Property="StrokeIconData" Value="M 4,0 L 4,2 7,2 7,0 Z M 4,4 L 4,10 7,10 7,4 Z"/>
+			<Setter Property="FillIconData" Value="M 4,0 L 4,2 7,2 7,0 Z M 7,10 7,4"/>
+		</Style>
+
 		<!-- Number Label -->
 		<Style x:Key="NumberLabelStyle" TargetType="{x:Type Label}">
 			<Setter Property="VerticalAlignment" Value="Center"/>
@@ -619,6 +624,7 @@
 							<RowDefinition Height="Auto"/>
 							<RowDefinition Height="Auto"/>
 							<RowDefinition Height="Auto"/>
+							<RowDefinition Height="Auto"/>
 						</Grid.RowDefinitions>
 
 						<Grid Grid.Row="0">
@@ -655,6 +661,13 @@
 													 IsEnabled="{Binding IsContrastSupported, Mode=OneWay}"
 													 Visibility="{Binding Settings.EnablesContrast, Mode=OneWay, Converter={StaticResource BooleanToVisibilityConverterKey}}"
 													 IsChecked="{Binding IsContrastChanging, Mode=TwoWay}"/>
+
+								<!-- Identity -->
+								<controls:IconButton x:Name="IdentityButton"
+													 Style="{StaticResource IdentityButtonStyle}"
+													 ToolTip="{x:Static properties:Resources.ShowIdentity}"
+													 Visibility="{Binding Settings.EnablesIdentity, Mode=OneWay, Converter={StaticResource BooleanToVisibilityConverterKey}}"
+													 IsChecked="{Binding IsIdentityShown, Mode=TwoWay}"/>
 							</StackPanel>
 						</Grid>
 
@@ -756,15 +769,38 @@
 								   Visibility="{Binding IsContrastChanging, Mode=OneWay, Converter={StaticResource BooleanToVisibilityConverterKey}}"/>
 						</Grid>
 
-						<!-- Message -->
+						<!-- Identity -->
 						<Grid Grid.Row="2"
+							  Visibility="{Binding IsIdentityShown, Mode=OneWay, Converter={StaticResource BooleanToVisibilityConverterKey}}">
+							<StackPanel Margin="12,2,12,2"
+										TextElement.Foreground="{StaticResource App.Foreground}"
+										TextElement.FontSize="10">
+								<TextBlock>
+									<Run Text="{Binding Edid.Manufacturer, Mode=OneTime, StringFormat='Manufacturer: {0}\,'}"/>
+									<Run Text="{Binding Edid.ManufactureYear, Mode=OneTime, StringFormat='Year: {0}\,'}"/>
+									<Run Text="{Binding Edid.SerialNumber, Mode=OneTime, StringFormat='S/N: {0}', TargetNullValue=''}"/>
+									<LineBreak/>
+									<Run Text="{Binding Connection, Mode=OneTime, StringFormat=Connection: {0}}"/>
+								</TextBlock>
+								<StackPanel Orientation="Horizontal">
+									<TextBlock VerticalAlignment="Center"
+											   Text="Device Instance ID:"/>
+									<controls:FocusTextBox VerticalAlignment="Center"
+														   Style="{StaticResource IdentityTextBoxStyle}"
+														   Text="{Binding DeviceInstanceId, Mode=OneTime}"/>
+								</StackPanel>
+							</StackPanel>
+						</Grid>
+
+						<!-- Message -->
+						<Grid Grid.Row="3"
 							  Visibility="{Binding NormalMessage, Mode=OneWay, Converter={StaticResource ObjectToVisibilityConverterKey}}">
 							<TextBlock Style="{StaticResource PlainTextStyle}"
 									   Margin="12,2,12,2"
 									   TextWrapping="Wrap"
 									   Text="{Binding NormalMessage, Mode=OneWay}"/>
 						</Grid>
-						<Grid Grid.Row="3"
+						<Grid Grid.Row="4"
 							  Visibility="{Binding WarningMessage, Mode=OneWay, Converter={StaticResource ObjectToVisibilityConverterKey}}">
 							<TextBlock Style="{StaticResource InvalidTextStyle}"
 									   Margin="12,2,12,2"
