+++ Source/Monitorian.Core/Views/MenuWindow.xaml
@@ -139,6 +139,13 @@
 						  IsChecked="{Binding Settings.EnablesContrast}"/>
 		</ContentControl>
 
+		<ContentControl Style="{StaticResource MenuItemStyle}">
+			<ToggleButton Padding="8,4"
+						  Style="{StaticResource CheckButtonItemStyle}"
+						  Content="{x:Static properties:Resources.EnableIdentity}"
+						  IsChecked="{Binding Settings.EnablesIdentity}"/>
+		</ContentControl>
+
 		<Separator Style="{StaticResource MenuSeparatorStyle}"/>
 
 		<!-- Close -->
